+++ src/main/java/com/getindata/connectors/http/internal/utils/ApiLookupConstants.java
@@ -1,0 +1,8 @@
+package com.getindata.connectors.http.internal.utils;
+
+public interface ApiLookupConstants {
+    String REQUEST_ARG_NAME_PREFIX = "request_";
+    String REQUEST_BODY_ARG_NAME_PREFIX = "requestBody_";
+    String REQUEST_PATH_ARG_NAME_PREFIX = "requestPath_";
+    String RESPONSE_ARG_NAME_PREFIX = "response_";
+}