+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+-   Added support for generic json and URL query creator 
+
 -   Retries support for source table:
     -   Auto retry on IOException and user-defined http codes - parameter `gid.connector.http.source.lookup.retry-codes`.
     -   Parameters `gid.connector.http.source.lookup.error.code.exclude"` and `gid.connector.http.source.lookup.error.code` were replaced by `gid.connector.http.source.lookup.ignored-response-codes`.
@@ -12,6 +14,8 @@
 -   OIDC token request to not flow during explain
 
 ## [0.18.0] - 2025-01-15
+
+### Fixed
 
 -   Ignore Eclipse files in .gitignore
 -   Support Flink 1.20